Unify importing of BOMS for different source file types Add support for multiple valid header names. Uses indices for each header vs. searching for every row. Add more valid headers. Remove test code
